.profile updates, more curly braces
For some reason we want to set ENV for Korn shells. But it's
not easy to find out if we are running a ksh version. KSH_VERSION
may not be set, and SHELL may not have been updated. And this
"ps -p" trick doesn't work on Busybox since "-p" is not understood.
So let's try an either/or approach here.

Signed-off-by: Christian Kujau <lists@nerdbynature.de>